test(calendar): migrate recurrence form-behavior tests to RecurrenceEditorSheet (pv-j1pi.4)
Audit of three source files:
- event-recurrence-init.test.ts: migrated endType + weekday checkbox init
  tests (5 blocks) to the sheet test file. Retained showRecurrence trigger
  tests (5 blocks) and useEventDuplication tests (2 blocks) with inline
  comments explaining why they stay.
- event-recurrence-dst.test.ts: all 3 blocks retained — they exercise
  date/time/timezone inputs which stay on event_recurrence.vue per the
  epic DESIGN (date/time inputs explicitly out of scope).
- event-recurrence-end-time.test.ts: all 6 blocks retained for the same
  reason — end-date/end-time inputs stay on the wrapper component.

The inline form in event_recurrence.vue is untouched; removal happens
in pv-j1pi.5. Coverage is preserved against the sheet before the inline
form is deleted (testing-advisor requirement).
